Week 7
This week, we worked on GUI panels. We made a message appear when getting close to a question mark and disappear when you moved away.

I also made a script that rotated the question mark and added a mathematical sinusoidal bounce to it so it looks fun.

I modified the trigger script for the mechanism of the balls to add an optional help mark that can be destroyed along with the triggers if the setting is enabled.