Week 8
Didn't do much this week. This week was all about finding assets and implementing them into our games. We were tasked to begin looking for assets for this class' assessment and for CI413's.

I put in a few buildings and added mesh colliders to them. I would do just multiple box colliders but I am lazy. I did do it on one of them but again: lazy.

Had to exclude most of the external assets we were given from the package because it was "too big", according to GitHub and it wouldn't be able to be published.

Soon, I will start working on my assessment work. I made the folder to hold all of it.